rename 'iranianNamesDataset.py to 'iranian_names_dataset.py'
