+++ persian_gender_detection/gender/iranian_names_dataset.py
